now it'll work
i was having a big problem with: (most likely due to a circular import)

it was happening because i imported the app directly from other parts of the code to catch the Database.

I fixed it by creating a separate file to instantiate the Database, then export to other files.